--- src/main/java/com/twitter_dan/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.twitter_dan.controller;
-
-public class UserController {
-}
