--- src/main/java/com/twitter_dan/dao/TweetsDao.java
@@ -1,4 +1,0 @@
-package com.twitter_dan.dao;
-
-public class TweetsDao {
-}
