--- src/main/java/com/twitter_dan/facade/TweetFacade.java
@@ -1,4 +1,0 @@
-package com.twitter_dan.facade;
-
-public class TweetFacade {
-}
