--- src/main/java/com/twitter_dan/service/LikeService.java
@@ -1,4 +1,0 @@
-package com.twitter_dan.service;
-
-public class LikeService {
-}
