--- src/main/java/com/twitter_dan/service/TweetService.java
@@ -1,4 +1,0 @@
-package com.twitter_dan.service;
-
-public class TweetService {
-}
