--- src/main/java/com/twitter_dan/service/UserService.java
@@ -1,4 +1,0 @@
-package com.twitter_dan.service;
-
-public class UserService {
-}
