+++ src/main/java/com/twitterdan/controller/ControllerAdvice.java
@@ -1,0 +1,4 @@
+package com.twitterdan.controller;
+
+public class ControllerAdvice {
+}
