+++ src/main/java/com/twitterdan/controller/LikesController.java
@@ -1,0 +1,4 @@
+package com.twitterdan.controller;
+
+public class LikesController {
+}
