+++ src/main/java/com/twitterdan/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.twitterdan.controller;
+
+public class UserController {
+}
