+++ src/main/java/com/twitterdan/dao/LikesDao.java
@@ -1,0 +1,4 @@
+package com.twitterdan.dao;
+
+public class LikesDao {
+}
