+++ src/main/java/com/twitterdan/dao/TweetsDao.java
@@ -1,0 +1,4 @@
+package com.twitterdan.dao;
+
+public class TweetsDao {
+}
