+++ src/main/java/com/twitterdan/facade/TweetFacade.java
@@ -1,0 +1,4 @@
+package com.twitterdan.facade;
+
+public class TweetFacade {
+}
