+++ src/main/java/com/twitterdan/facade/UserFacade.java
@@ -1,0 +1,4 @@
+package com.twitterdan.facade;
+
+public class UserFacade {
+}
