+++ src/main/java/com/twitterdan/service/UserService.java
@@ -1,0 +1,4 @@
+package com.twitterdan.service;
+
+public class UserService {
+}
