+++ src/main/resources/application.yml
@@ -25,7 +25,7 @@ spring:
     url: jdbc:h2:mem:testdb
     username: sa
     password:
-    driver-class-name: org.h2.driver
+    driver-class-name: org.h2.Driver
   jpa:
     defer-datasource-initialization: true
     properties:
