+++ src/test/java/com/twitterdan/FsOnlineFinalProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package com.twitter.danit;
+package com.twitterdan;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
